Rename xesc2_mini hardware target to xesc_all_variants
Makes it clear the firmware supports all xESC variants (mini, power, lite),
not just the mini. Renames directory, hw_xesc2_mini.{h,c} files, updates
include, hwconf.mk paths, and CI workflow target/artifact names.
