+++ .github/workflows/build.yml
@@ -23,16 +23,16 @@ jobs:
         if: steps.cache-arm-sdk.outputs.cache-hit != 'true'
         run: make arm_sdk_install
 
-      - name: Build xesc2_mini
-        run: make xesc2_mini
+      - name: Build xesc_all_variants
+        run: make xesc_all_variants
 
       - name: Upload artifacts
         uses: actions/upload-artifact@v4
         with:
           name: firmware-${{ github.sha }}
           path: |
-            build/xesc2_mini/xesc2_mini.bin
-            build/xesc2_mini/xesc2_mini.elf
+            build/xesc_all_variants/xesc_all_variants.bin
+            build/xesc_all_variants/xesc_all_variants.elf
 
   release:
     if: startsWith(github.ref, 'refs/tags/')
@@ -52,5 +52,5 @@ jobs:
         uses: softprops/action-gh-release@v2
         with:
           files: |
-            artifacts/xesc2_mini.bin
-            artifacts/xesc2_mini.elf
+            artifacts/xesc_all_variants.bin
+            artifacts/xesc_all_variants.elf
