+++ hwconf/xtech/xesc_all_variants/hw_xesc_all_variants.h
