+++ hwconf/xtech/xesc_all_variants/xesc2_otp.h
