+++ hwconf/xtech/xesc_all_variants/xesc2_variant_config.h
